+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "marketplace",
   "version": "0.1.0",
-  "homepage": "https://artem-andreevich.github.io/marketplace/",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.17.0",
