+++ src/hoc/Layout.tsx
@@ -1,9 +1,7 @@
-import React from 'react'
-import Header from './Header/Header'
-import Footer from './Footer/Footer'
+import { Header, Footer} from "../layout"
 import { Outlet } from 'react-router-dom'
 
-const Layout = () => {
+export const Layout = () => {
     return (
         <div className='wrapper'>
             <Header />
@@ -11,6 +9,4 @@ const Layout = () => {
             <Footer />
         </div>
     )
-}
-
-export default Layout
+};
