+++ src/layout/Footer/Footer.tsx
@@ -1,6 +1,4 @@
-import React from 'react';
-
-const Footer = () => {
+export const Footer = () => {
     return (
         <footer className='footer'>
             <div className='container'>
@@ -9,5 +7,3 @@ const Footer = () => {
         </footer>
     );
 };
-
-export default Footer;
