+++ src/layout/Header/index.js
@@ -1,0 +1,1 @@
+export { Header } from "./Header"
