+++ src/layout/Portal/index.js
@@ -1,0 +1,1 @@
+export { Svg } from "./Svg"
