+++ src/layout/index.js
@@ -1,0 +1,3 @@
+export * from "./Header";
+export * from "./Footer";
+export * from "./Portal";
