+++ src/pages/index.js
@@ -1,0 +1,5 @@
+export * from "./AboutPage";
+export * from "./CategoriesInnerPage";
+export * from "./CategoriesPage";
+export * from "./MainPage";
+export * from "./ProductPage";
